--- src/app/(home)/layout.js
@@ -1,5 +1,0 @@
-import AppLayout from "../../components/layout/AppLayout";
-
-export default function Layout({ children }) {
-  return <AppLayout fadeNavbar={true}>{children}</AppLayout>;
-}
