+++ src/components/layout/AppLayout.jsx
@@ -9,10 +9,10 @@ import AppFooter from "@/components/layout/AppFooter";
  */
 export default function AppLayout({ children, fadeNavbar = false }) {
   return (
-    <div className="">
+    <>
       <AppNavbar fade={fadeNavbar} />
       <div className="toolbar-after-content">{children}</div>
       <AppFooter />
-    </div>
+    </>
   );
 }
